Replace `zip-folder` to resolve a security warning
`zip-folder` has not been updated to track major version bumps in its
dependency `archive`, and it does not publish a version without that
module's v0.11 vulnerabilities.

Since `zip-folder` is a small wrapper around `archive`, the easiest fix is
to provide the feature directly on the current `archive` module.

This commit replaces `zip-folder` with a direct dependency on `archive`
and adds a `zipFolder` function that replicates the `zip-folder` feature.