Fix cache misses with duplicate requests
The cache is an array of pairs containing the parameter object and first
SandTable Server response. The former is the key to be compared against
for future requests. If a request matches the key of another, it will
re-use the response (namely the request ID) of the previous response.

It has to be done in this ad-hoc manner as JavaScript cannot use deep
value comparison for keys in objects/Maps. This results in the loss of
the O(1) performance of a proper look-up table, however this should not
be too large an issue.